--- src/db/drop_all_tables.sql
@@ -1,15 +1,0 @@
-DROP TABLE PlayersBasicInfo;
-DROP TABLE Positions;
-DROP TABLE Playstyles;
-DROP TABLE Versions;
-DROP TABLE Nations;
-DROP TABLE Leagues;
-DROP TABLE Clubs;
-DROP TABLE LeagueClubs;
-DROP TABLE AcceleRATE;
-DROP TABLE Players;
-DROP TABLE PlayerAltPositions;
-DROP TABLE PlayerPlaystyles;
-DROP TABLE PlayerPlaystylesPlus;
-DROP TABLE NationalityPlayers;
-DROP TABLE ClubsPlayers;
