+++ src/utils/csv_utils.py
@@ -2,7 +2,6 @@ import pandas as pd
 
 from futwiz.player_page.player_data_template import PlayerDataTemplateFactory
 from ut_players.common.utils import does_file_include_player_stats
-from enum import Enum
 
 
 def get_csv_content(filepath, delimiter=';'):
